Add PLTADDONDIR environment variable to racket binaries.
The PLTADDONDIR contains user-specific Racket configuration:
https://docs.racket-lang.org/reference/Filesystem.html#%28idx._%28gentag._356._%28lib._scribblings%2Freference%2Freference..scrbl%29%29%29

For example, is used by `raco` to install and locate a user’s `raco-commands`.

For the builder, it defaults to the user’s home directory
`$TMP_RACO_HOME`. The built `raco` binary, however, defaults to the
homeless shelter. This means that the built `raco` can't locate any
`raco-commands` in packages installed by the builder.

This commit creates a PLTADDONDIR directory structure in `etc` and
sets it as an environment variable to all binaries. The paths in
`config.rkdt` also point to this directory structure.