feat(lock): re-implement in python...
as discussed during the last weekly, where we've implemented a
prototype, this replaces the nix + jq magic with python.

it requires more SLOCs and error handling could further be improved,
but i think it helps readability and makes the logic of parsing
nix build logs a bit more robust by checking that we get a hash
for the drv_path we expected.

Empty hash is replaced by l.fakeSha256 to silence a nix warning.