v1/pip: support building with a subset of locked extras (#530)
* fix(v1/pip): remove substitution opt...

...from the interface, as it's already gone in the implementation.

There's no direct replacement for binary substitutions anymore, use
nixpkgs-overrides as apropriate.

* feat(v1/pip): default to provided_extras for lock

* feat(v1/pip, breaking): introduce targets to lock.

...targets correspond to a set of python extra markers, e.g. "dev"
or "docs". We now write separate dependency trees for each requested
extra to the lock file.

We can then select a subset of those features using the new
"pip.buildExtras" option. Note that this option currently affects the
dependency tree only.

This approach is still not bullet proof, i.e. in cases where a
constraint of a package in "dev" changes the version of a package
in "default" during resolving.