feat(nodejs-node-modules): ensure the .bin also exists in the local node_modules
This change allows to re-locate the node_modules directory via a single copy operation.

The current builder installs the .bin directory one level above the top-level package which isn't what we want for the node-modules output.

This change ensures that the .bin also exists within the local node_modules directory of the current package.

It also ensures that all symlinks are relative and don't point to store paths