pip: don't crash on cycles, fix them instead
pip does allow cycles and they are common in larger dependency trees. Instead of crashing on them, we should allow the user to deal with them.

As a result of this, it can happen that during the installation of individual dependencies, some dep(s) might be missing, which then can be fixed by including them manually or with a `--no-deps`