package-func: select first output in outputs for multi-output derivations
This commit modifies the package-func module to use the first output in
`package-func.outputs` or the default output (if the attribute `outputSpecified`
is true) as the top-level output in the final package instead of throwing
an error for multi-output derivations. Documentation for structure of the
assembled package and the selection criterion for the top-level output added
to the README.md under the module directory. In addition, a changelog entry
was appended to CHANGELOG.md to record this change.